Minor: Add comment. Do not typecast u16 as u16

+++ src/cpu/instructions/lda.rs
@@ -55,7 +55,7 @@ pub(super) fn immediate(arg: &mut CpuInstructionFnArguments) -> Option<u8> {
     match arg.cpu.registers.get_flag(StatusFlags::AccSize) {
         REGISTER_MODE_8_BIT => {
             let masked_param = arg.param & 0x00FF;
-            arg.cpu.registers.acc = Wrapping(masked_param as u16);
+            arg.cpu.registers.acc = Wrapping(masked_param);
         }
         REGISTER_MODE_16_BIT => {
             arg.cpu.registers.acc = Wrapping(arg.param);
@@ -144,9 +144,7 @@ pub(super) fn absolute_long(arg: &mut CpuInstructionFnArguments) -> Option<u8> {
 ///
 /// Returns:
 ///     - `Some(3)`: Number of cycles to pend.
-pub(super) fn direct_page(arg: &mut CpuInstructionFnArguments) -> Option<u8> {
-    Some(3)
-}
+pub(super) fn direct_page(arg: &mut CpuInstructionFnArguments) -> Option<u8> { Some(3) }
 
 /**************************************** Tests *************************************************************************/
